+++ java-client/src/main/java/org/opensearch/client/opensearch/indices/stats/ShardStats.java
@@ -110,6 +110,7 @@ public class ShardStats implements PlainJsonSerializable {
     @Nullable
     private final BulkStats bulk;
 
+    @Nullable
     private final ShardsTotalStats shards;
 
     // ---------------------------------------------------------------------------------------------
@@ -138,8 +139,7 @@ public class ShardStats implements PlainJsonSerializable {
         this.translog = ApiTypeHelper.requireNonNull(builder.translog, this, "translog");
         this.warmer = ApiTypeHelper.requireNonNull(builder.warmer, this, "warmer");
         this.bulk = builder.bulk;
-        this.shards = ApiTypeHelper.requireNonNull(builder.shards, this, "shards");
-
+        this.shards = builder.shards;
     }
 
     public static ShardStats of(Function<Builder, ObjectBuilder<ShardStats>> fn) {
@@ -302,8 +302,9 @@ public class ShardStats implements PlainJsonSerializable {
     }
 
     /**
-     * Required - API name: {@code shards}
+     * API name: {@code shards}
      */
+    @Nullable
     public final ShardsTotalStats shards() {
         return this.shards;
     }
@@ -387,9 +388,10 @@ public class ShardStats implements PlainJsonSerializable {
             this.bulk.serialize(generator, mapper);
 
         }
-        generator.writeKey("shards");
-        this.shards.serialize(generator, mapper);
-
+        if (this.shards != null) {
+            generator.writeKey("shards");
+            this.shards.serialize(generator, mapper);
+        }
     }
 
     // ---------------------------------------------------------------------------------------------
@@ -444,6 +446,7 @@ public class ShardStats implements PlainJsonSerializable {
         @Nullable
         private BulkStats bulk;
 
+        @Nullable
         private ShardsTotalStats shards;
 
         /**
@@ -777,7 +780,7 @@ public class ShardStats implements PlainJsonSerializable {
         }
 
         /**
-         * Required - API name: {@code shards}
+         * API name: {@code shards}
          */
         public final Builder shards(ShardsTotalStats value) {
             this.shards = value;
@@ -785,7 +788,7 @@ public class ShardStats implements PlainJsonSerializable {
         }
 
         /**
-         * Required - API name: {@code shards}
+         * API name: {@code shards}
          */
         public final Builder shards(Function<ShardsTotalStats.Builder, ObjectBuilder<ShardsTotalStats>> fn) {
             return this.shards(fn.apply(new ShardsTotalStats.Builder()).build());